Refactor: Move parts of the address poisoning detection logic to a lib to keep the controller leaner
